add ability to render assignments posted by teacher
Now, both teacherController and studentController can handle get requests to courses/course_code/, which renders the assignments.
The Assignment and Question model also have changes to make this possible.
Notably, the assignmentSchema now has a query middleware to populate the questions, and are sorted by createdAt Date, to ensure correct numbering when rendered in views.